rbi-assessment and select list
