+++ src/pages/maintain/pm-schedule/PMScheduleDetailPage.tsx
@@ -187,11 +187,14 @@ const PMScheduleDetailPage: React.FC = () => {
               <TabsContent value="checksheet">
                 {id && <ChecksheetTab pmScheduleId={Number(pmSchedule.id)} />}
               </TabsContent>
-              {id && pmSchedule.asset_id?.id ? (
-                <RelatedWoTab assetId={Number(pmSchedule.asset_id.id)} />
-              ) : (
-                <div>No related work orders available for this asset.</div>
-              )}
+              <TabsContent value="workOrder">
+                {id && pmSchedule.asset_id?.id ? (
+                  <RelatedWoTab assetId={Number(pmSchedule.asset_id.id)} />
+                ) : (
+                  <div>No related work orders available for this asset.</div>
+                )}
+              </TabsContent>
+
               <TabsContent value="additionalInfo">
                 {id && <AdditionalInfoTab pmScheduleId={Number(pmSchedule.id)} />}
               </TabsContent>
