+++ src/routes/AppRoutes.tsx
@@ -105,8 +105,8 @@ import NewPressureVesselPage from "@/pages/monitor/NewPressureVesselPage";
 import NewPipingPage from "@/pages/monitor/NewPipingPage";
 import FormulaTestingPage from "@/pages/monitor/FormulaTestingPage";
 import InventoryGroupsPage from "@/pages/monitor/InventoryGroupsPage";
-import RBIAssessmentPage from "@/pages/monitor/RBIAssessmentPage";
-import RBIAssessmentDetailPage from "@/pages/monitor/RBIAssessmentDetailPage";
+import RBIAssessmentPage from "@/pages/monitor/rbi/RBIAssessmentPage";
+import RBIAssessmentDetailPage from "@/pages/monitor/rbi/RBIAssessmentDetailPage";
 // import RBIAssessmentPVDetailPage from "@/pages/monitor/RBIAssessmentPVDetailPage";
 import RMSAssetListPage from "@/pages/monitor/RMSAssetListPage";
 import RMSAssetDetailPage from "@/pages/monitor/RMSAssetDetailPage";
